Create Q#6.cpp

+++ ETS1303_15_TESFAMICHAEL_TAFERE/Q#6.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 int main()
 {
-    const double transmmitionRate =960;
+    const float transmmitionRate =960;
     int fileSize;
     float time, hour;
